Don't include day input in livemd file
That was ok-ish in the `setup` section of a livebook but when viewed as
markdown it meant having to scroll for days to reach the actual code.

I've made getting to the content as painless as possible by linking to
the advent of code input and to my input in this repo as txt.